+++ types/lodash.iserror/index.d.ts
@@ -1,2 +1,0 @@
-declare function isError(obj: any): obj is Error;
-export = isError;